requirements.txt
apache-airflow>=2.5.0
scikit-learn>=0.24.0
numpy>=1.19.0